+++ packages/hash/types/index.d.ts
@@ -1,3 +1,3 @@
 import { Options, HashHistory } from "./types";
 export * from "./types";
-export declare function Hash<Q>(options?: Options<Q>): HashHistory<Q>;
+export declare function Hash(options?: Options): HashHistory;