+++ packages/root/types/prepareNavigate.d.ts
@@ -1,2 +1,2 @@
 import { PrepNavigateArgs, Preparer } from "./types/prepareNavigate";
-export default function prepareNavigate<Q>(args: PrepNavigateArgs<Q>): Preparer<Q>;
+export default function prepareNavigate(args: PrepNavigateArgs): Preparer;