+++ packages/root/types/types/navigation.d.ts
@@ -1,12 +1,12 @@
 import { PartialLocation, SessionLocation } from "./location";
-export declare type ToArgument<Q> = string | PartialLocation<Q>;
+export declare type ToArgument = string | PartialLocation;
 export declare type Action = "push" | "replace" | "pop";
 export declare type NavType = "anchor" | "push" | "replace";
-export interface PendingNavigation<Q> {
-    location: SessionLocation<Q>;
+export interface PendingNavigation {
+    location: SessionLocation;
     action: Action;
     finish(): void;
     cancel(nextAction?: Action): void;
     cancelled?: boolean;
 }
-export declare type ResponseHandler<Q> = (resp: PendingNavigation<Q>) => void;
+export declare type ResponseHandler = (resp: PendingNavigation) => void;